requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.2
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
Werkzeug==1.0.1
